+++ stroom-core-server/src/main/java/stroom/process/server/ExpressionToFindStreamProcessorFilterCriteria.java
@@ -6,7 +6,6 @@ import stroom.entity.shared.BaseEntity;
 import stroom.entity.shared.EntityIdSet;
 import stroom.entity.shared.EntityServiceException;
 import stroom.entity.shared.StringCriteria;
-import stroom.explorer.server.ExplorerService;
 import stroom.pipeline.server.PipelineService;
 import stroom.pipeline.shared.FindPipelineEntityCriteria;
 import stroom.pipeline.shared.PipelineEntity;
@@ -15,6 +14,7 @@ import stroom.query.api.v2.ExpressionItem;
 import stroom.query.api.v2.ExpressionOperator;
 import stroom.query.api.v2.ExpressionOperator.Op;
 import stroom.query.api.v2.ExpressionTerm;
+import stroom.streamstore.server.CollectionService;
 import stroom.streamstore.shared.QueryData;
 import stroom.streamstore.shared.StreamDataSource;
 import stroom.streamtask.shared.FindStreamProcessorFilterCriteria;
@@ -37,7 +37,7 @@ import java.util.stream.Collectors;
 public class ExpressionToFindStreamProcessorFilterCriteria {
     private final PipelineService pipelineService;
     private final DictionaryStore dictionaryStore;
-    private final ExplorerService explorerService;
+    private final CollectionService collectionService;
 
     private static final Function<List<Op>, String> OP_STACK_DISPLAY = (s) ->
             s.stream().map(Op::getDisplayValue).collect(Collectors.joining(" -> "));
@@ -48,10 +48,10 @@ public class ExpressionToFindStreamProcessorFilterCriteria {
     @Inject
     public ExpressionToFindStreamProcessorFilterCriteria(@Named("cachedPipelineService") final PipelineService pipelineService,
                                                          final DictionaryStore dictionaryStore,
-                                                         final ExplorerService explorerService) {
+                                                         final CollectionService collectionService) {
         this.pipelineService = pipelineService;
         this.dictionaryStore = dictionaryStore;
-        this.explorerService = explorerService;
+        this.collectionService = collectionService;
     }
 
     public FindStreamProcessorFilterCriteria convert(final ExpressionOperator findStreamCriteria) {
@@ -471,7 +471,7 @@ public class ExpressionToFindStreamProcessorFilterCriteria {
                     values.addAll(words);
                     break;
                 case IN_FOLDER:
-                    final Set<DocRef> descendants = explorerService.getDescendants(term.getDocRef(), term.getField());
+                    final Set<DocRef> descendants = collectionService.getDescendants(term.getDocRef(), term.getField());
                     values.addAll(descendants.stream().map(DocRef::getUuid).collect(Collectors.toSet()));
                     break;
                 case IS_DOC_REF: