+++ stroom-core-server/src/main/java/stroom/streamstore/server/ExpressionMatcherFactory.java
@@ -3,7 +3,6 @@ package stroom.streamstore.server;
 import org.springframework.stereotype.Component;
 import stroom.datasource.api.v2.DataSourceField;
 import stroom.dictionary.server.DictionaryStore;
-import stroom.explorer.server.ExplorerService;
 
 import javax.inject.Inject;
 import java.util.Map;
@@ -11,15 +10,15 @@ import java.util.Map;
 @Component
 public class ExpressionMatcherFactory {
     private final DictionaryStore dictionaryStore;
-    private final ExplorerService explorerService;
+    private final CollectionService collectionService;
 
     @Inject
-    public ExpressionMatcherFactory(final DictionaryStore dictionaryStore, final ExplorerService explorerService) {
+    public ExpressionMatcherFactory(final DictionaryStore dictionaryStore, final CollectionService collectionService) {
         this.dictionaryStore = dictionaryStore;
-        this.explorerService = explorerService;
+        this.collectionService = collectionService;
     }
 
     public ExpressionMatcher create(final Map<String, DataSourceField> fieldMap) {
-        return new ExpressionMatcher(fieldMap, dictionaryStore, explorerService);
+        return new ExpressionMatcher(fieldMap, dictionaryStore, collectionService);
     }
 }